added JWT

+++ bitirmeprojesi-ArdaNogay/src/main/java/com/softtech/softtechspringboot/Dto/CategorySaveAndUpdateRequestDto.java
@@ -3,9 +3,11 @@ package com.softtech.softtechspringboot.Dto;
 import com.softtech.softtechspringboot.Enum.CategoryType;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CategorySaveAndUpdateRequestDto {
 
     private CategoryType categoryType;
-    private Double tax; //Todo: BigDecimal yapılacak burası
+    private BigDecimal tax; //Todo: BigDecimal yapılacak burası
 }
