+++ bitirmeprojesi-ArdaNogay/src/main/java/com/softtech/softtechspringboot/Dto/ProductSaveAndUpdateRequestDto.java
@@ -2,11 +2,13 @@ package com.softtech.softtechspringboot.Dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 
 @Data
 public class ProductSaveAndUpdateRequestDto {
 
     private String name;
-    private Double taxFreePrice;
+    private BigDecimal taxFreePrice;
     private Long categoryId;
 }
