+++ bitirmeprojesi-ArdaNogay/src/main/java/com/softtech/softtechspringboot/Entity/Category.java
@@ -24,6 +24,6 @@ public class Category extends BaseEntity {
     private CategoryType categoryType;
 
     @Column(name = "TAX" ,nullable = false ,updatable = true)
-    private Double tax; //Todo: BigDecimal yapılacak burası
+    private BigDecimal tax; //Todo: BigDecimal yapılacak burası
 
 }
